// src/app/core/models/enums/date-time-locale-format.enum.ts
export enum DateTimeLocaleFormat {
    deDisplayShort = 'dd.MM.yyyy',
    deDisplayLong = 'dd.MM.yyyy HH:mm',
    deMomentDisplayShort = 'DD.MM.yyyy',
    deMomentDisplayLong = 'DD.MM.yyyy HH:mm',
    deDisplayShortPlaceholder = 'tt.mm.jjjj',
    enDisplayShort = 'MM/dd/yyyy',
    enDisplayLong = 'MM/dd/yyyy HH:mm',
    enMomentDisplayShort = 'MM/DD/yyyy',
    enMomentDisplayLong = 'MM/DD/yyyy HH:mm',
    enDisplayShortPlaceholder = 'mm/dd/yyyy',
}